My emacs is just too clever for its own good...
git-svn-id: svn://svn.valgrind.org/valgrind/trunk@5030

+++ coregrind/launcher.c
@@ -105,7 +105,7 @@ static const char *select_platform(const char *clientname)
    int fd;
    unsigned char *header;
    const char *platform = NULL;
-   int32_t pagesize = sysconf(_SC_PAGESIZE);
+   long pagesize = sysconf(_SC_PAGESIZE);
 
    if (strchr(clientname, '/') == NULL)
       clientname = find_client(clientname);
